+++ extensions/iceberg/src/test/java/io/deephaven/iceberg/internal/ResolverFromTest.java
@@ -47,7 +47,7 @@ class ResolverFromTest {
             Resolver.from(TableDefinition.of(ColumnDefinition.of("Foo", Type.shortType())));
             failBecauseExceptionWasNotThrown(Resolver.MappingException.class);
         } catch (Resolver.MappingException e) {
-            assertThat(e).hasMessageContaining("Unsupported deephaven column type io.deephaven.qst.type.ShortType");
+            assertThat(e).hasMessageContaining("Unsupported Deephaven column type io.deephaven.qst.type.ShortType");
         }
     }
 
@@ -69,7 +69,7 @@ class ResolverFromTest {
             failBecauseExceptionWasNotThrown(Resolver.MappingException.class);
         } catch (Resolver.MappingException e) {
             assertThat(e).hasMessageContaining(
-                    "Unsupported deephaven column type NativeArrayType{clazz=class [B, componentType=io.deephaven.qst.type.ByteType}");
+                    "Unsupported Deephaven column type NativeArrayType{clazz=class [B, componentType=io.deephaven.qst.type.ByteType}");
         }
     }
 
@@ -80,7 +80,7 @@ class ResolverFromTest {
             failBecauseExceptionWasNotThrown(Resolver.MappingException.class);
         } catch (Resolver.MappingException e) {
             assertThat(e).hasMessageContaining(
-                    "Unsupported deephaven column type NativeArrayType{clazz=class [S, componentType=io.deephaven.qst.type.ShortType}");
+                    "Unsupported Deephaven column type NativeArrayType{clazz=class [S, componentType=io.deephaven.qst.type.ShortType}");
         }
     }
 
@@ -91,7 +91,7 @@ class ResolverFromTest {
             failBecauseExceptionWasNotThrown(Resolver.MappingException.class);
         } catch (Resolver.MappingException e) {
             assertThat(e).hasMessageContaining(
-                    "Unsupported deephaven column type NativeArrayType{clazz=class [I, componentType=io.deephaven.qst.type.IntType}");
+                    "Unsupported Deephaven column type NativeArrayType{clazz=class [I, componentType=io.deephaven.qst.type.IntType}");
         }
     }
 }
